backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend"
authors = ["samiribrh <samiribrahimov2277@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "2.3.3"
flask-httpauth = "4.8.0"
flask-jwt-extended = "4.6.0"
itsdangerous = "^2.2.0"
mysql-connector-python = "8.1.0"
mysqlclient = "2.2.4"
pydantic = "^2.8.2"
pydantic-core = "^2.20.1"
python-dateutil = "^2.9.0.post0"
six = "^1.16.0"
sqlalchemy = "^2.0.31"
werkzeug = "^3.0.3"
python-dotenv = "^1.0.1"
flake8 = "^7.1.0"
bcrypt = "^4.2.0"
requests = "^2.32.3"
flask-swagger-ui = "^4.11.1"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
